safari-534.53-branch/LayoutTests/media/controls-right-click-on-timebar-expected.txt

Test that right clicking on the timebar does not cause a seek.

RUN(video.autoplay = true)
EVENT(playing)
END OF TEST
